✨ *refact(celery_app): melhorar a lógica de execução e manipulação de tarefas*
Refatora o módulo `celery_app` para otimizar a execução de tarefas e a manipulação de logs. As alterações incluem a mudança do pool de execução do worker para "solo" e a melhoria na lógica de manipulação de arquivos de log.

Files Changed:
- *celery_app/main.py*: altera o pool do worker para "solo" e ajusta a lógica de manipulação de logs.
- *celery_app/tasks/bot.py*: adiciona um getter e setter para o atributo `pid`, melhora a lógica de impressão de mensagens e remove código desnecessário.
- *celery_app/tasks/files.py*: renomeia a tarefa de salvar cache e simplifica a lógica de salvamento.
- *crawjud/bots/pje/movimentacao.py*: refatora a lógica de busca de processos e melhora a manipulação de erros.
- *crawjud/bots/pje/res/buscador.py*: substitui a dependência de `AsyncPrintMessage` por uma função de impressão genérica.
- *crawjud/core/__init__.py*: remove a classe `PropertiesCrawJUD` e integra suas propriedades na classe `CrawJUD`.
- *models/logs.py*: adiciona um novo campo `processo` na classe `CachedExecution`.
