+++ crawjud/_wrapper.py
@@ -1,21 +1,23 @@
-from typing import AnyStr, Callable, ParamSpec, TypeVar
+from functools import wraps
+from typing import AnyStr, ParamSpec, TypeVar
 
 from common.bot import ClassBot
 
-TClassBot = TypeVar("TClassBot", bound=ClassBot)
+TClassBot = TypeVar("TClassBot", bound=object)
 TBotSpec = ParamSpec("TBotSpec", bound=AnyStr)
 
 
-def wrap_init(cls: type[TClassBot]) -> type[TClassBot]:  # noqa: D103
-    original_init: Callable[TBotSpec, None] = cls.__init__
+def wrap_init(cls: type[ClassBot]) -> type[TClassBot]:  # noqa: D103
+    original_init = cls.__init__
 
+    @wraps(original_init)
     def novo_init(
         self: TClassBot,
         *args: TBotSpec.args,
         **kwargs: TBotSpec.kwargs,
     ) -> None:
         print(f"Instanciando {cls.__name__} com args: {args}, kwargs: {kwargs}")
-        original_init(self, *args, **kwargs)
+        original_init(self)
 
     cls.__init__ = novo_init
     return cls
